Add files via upload
This is the module for the camera function. I'll later post dependencies and the hardware required to use it. This is a simple but stand alone function for taking pictures and saving them to the desktop of a Raspberry Pi. Only works with Raspberry Pi and the Camera Module (regular issue).